Travis: Only build for latest stable releases (2.3.1 and 2.2.5)
Since this gem hasn't really been updated for a while, there are still builds listed for old, EOLed versions of Ruby. (1.8, 1.9 are all no longer maintained)

This commit removes those old builds and adds 2.2.5 and 2.3.1, while keeping the jruby-head and ruby-head builds.

This commit also resolves a style error with no newline at the end of the file.